Add a default implementation of Backend::print_at_rep and let the blt backend rely on it.
