+++ src/backend/mod.rs
@@ -10,6 +10,7 @@
 use crate::event::Event;
 use crate::theme;
 use crate::vec::Vec2;
+use unicode_width::UnicodeWidthStr;
 
 #[cfg(unix)]
 mod resize;
@@ -66,7 +67,21 @@ pub trait Backend {
     
     /// First positions the cursor, similar to `print_at`, and then prints the given number of
     /// `repetitions` of `text`.
-    fn print_at_rep(&self, pos: Vec2, repetitions: usize, text: &str);
+    fn print_at_rep(&self, pos: Vec2, repetitions: usize, text: &str) {
+        if repetitions > 0 {
+            self.print_at(pos, text);
+
+            let width = text.width();
+            let mut pos = pos;
+            let mut dupes_left = repetitions - 1;
+
+            while dupes_left > 0 {
+                pos = pos.saturating_add((width, 0));
+                self.print_at(pos, text);
+                dupes_left -= 1;
+            }
+        }
+    }
 
     /// Clears the screen with the given color.
     fn clear(&self, color: theme::Color);
